test/TwoToOneJavaTests.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

import org.junit.Test;
import static org.junit.Assert.*;
import twotoonejava.TwoToOneJava;

/**
 *
 * @author Mike Brooks
 */
public class TwoToOneJavaTests {
    
    @Test
    public void test() {
        System.out.println("longest Fixed Tests");
        assertEquals("aehrsty", TwoToOneJava.longest("aretheyhere", "yestheyarehere"));
        assertEquals("abcdefghilnoprstu", TwoToOneJava.longest("loopingisfunbutdangerous", "lessdangerousthancoding"));
        assertEquals("acefghilmnoprstuy", TwoToOneJava.longest("inmanylanguages", "theresapairoffunctions"));
    }
    
}
